requirements.txt
absl-py==0.9.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
backcall==0.1.0
beautifulsoup4==4.12.3
bleach==6.0.0
blender==1.4
cachetools==4.0.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
comm==0.1.4
cycler==0.10.0
debugpy==1.7.0
decorator==4.4.1
defusedxml==0.7.1
einops==0.6.1
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.1
filelock==3.12.2
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
gpytoolbox==0.2.0
grpcio==1.26.0
h5py==3.8.0
huggingface-hub==0.16.4
idna==2.8
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.16.0
Jinja2==3.1.3
joblib==0.14.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==4.12.0
jupyter-server==1.24.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==2.1.4
matplotlib==3.2.1
matplotlib-inline==0.1.6
mistune==3.0.2
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.4
notebook==6.5.6
notebook-shim==0.2.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
packaging==23.2
pandocfilters==1.5.1
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
plyfile==0.7.1
portalocker==2.7.0
progressbar==2.5
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==3.11.2
psutil==5.9.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.4.10
Pygments==2.5.2
pyparsing==2.4.6
pyquaternion==0.9.5
pyrsistent==0.19.3
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
regex==2023.10.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.8.2
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.4.1
tensorboard==2.0.2
tensorboardX==2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm
traitlets==5.4.0
trimesh==3.5.16
typing-extensions==4.7.1
urllib3==1.25.8
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==0.16.1
widgetsnbextension==4.0.9
zipp==3.15.0
